requirements.txt
pydantic==2.7.3
pyyaml==6.0.1
requests==2.32.3